[skip ci] go: use uint32 instead of uint16 for FAT32 support

+++ go/lib/fat/fat16/fat16.go
@@ -77,41 +77,41 @@ func NewFromPath(path string) (fat.FileSystem, error) {
 }
 
 const (
-	fatEntrySize = uint16(2)
-	eoc          = uint16(0xfff8)
+	fatEntrySize = uint32(2)
+	eoc          = uint32(0xfff8)
 )
 
 type fat16Table struct {
-	fatId      uint16
-	clusters   []uint16
-	maxCluster uint16
+	fatId      uint32
+	clusters   []uint32
+	maxCluster uint32
 }
 
 func parseFat16Table(fatBytes []byte) internal.FatTable {
-	maxCluster := uint16(len(fatBytes)) / fatEntrySize
+	maxCluster := uint32(len(fatBytes)) / fatEntrySize
 	fatTable := fat16Table{
-		fatId:      binary.LittleEndian.Uint16(fatBytes[0:fatEntrySize]),
-		clusters:   make([]uint16, maxCluster+1),
+		fatId:      uint32(binary.LittleEndian.Uint16(fatBytes[0:fatEntrySize])),
+		clusters:   make([]uint32, maxCluster+1),
 		maxCluster: maxCluster,
 	}
 
-	for cluster := uint16(2); cluster < maxCluster; cluster++ {
+	for cluster := uint32(2); cluster < maxCluster; cluster++ {
 		start := cluster * fatEntrySize
 		end := start + fatEntrySize
 		val := binary.LittleEndian.Uint16(fatBytes[start:end])
 		if val != 0 {
-			fatTable.clusters[cluster] = val
+			fatTable.clusters[cluster] = uint32(val)
 		}
 	}
 
 	return &fatTable
 }
 
-func (t *fat16Table) IsEoc(cluster uint16) bool              { return cluster >= eoc }
-func (t *fat16Table) GetEoc() uint16                         { return eoc }
-func (t *fat16Table) GetMaxCluster() uint16                  { return t.maxCluster }
-func (t *fat16Table) GetClusterTarget(cluster uint16) uint16 { return t.clusters[cluster] }
-func (t *fat16Table) WriteClusterTarget(fs *internal.FileSystem, cluster, target uint16) error {
+func (t *fat16Table) IsEoc(cluster uint32) bool              { return cluster >= eoc }
+func (t *fat16Table) GetEoc() uint32                         { return eoc }
+func (t *fat16Table) GetMaxCluster() uint32                  { return t.maxCluster }
+func (t *fat16Table) GetClusterTarget(cluster uint32) uint32 { return t.clusters[cluster] }
+func (t *fat16Table) WriteClusterTarget(fs *internal.FileSystem, cluster, target uint32) error {
 	// set in memory cluster table
 	t.clusters[cluster] = target
 
@@ -120,7 +120,7 @@ func (t *fat16Table) WriteClusterTarget(fs *internal.FileSystem, cluster, target
 		offset := fs.GetFatSectorOffset(i)
 		clusterOffset := offset + int64(cluster*fatEntrySize)
 		toWrite := make([]byte, 2)
-		binary.LittleEndian.PutUint16(toWrite, target)
+		binary.LittleEndian.PutUint16(toWrite, uint16(target))
 		_, err := fs.Backend.WriteAt(toWrite, clusterOffset)
 		if err != nil {
 			return err
